code cleaned
