+++ .gitignore
@@ -1,4 +1,4 @@
-images/
+diagrams/
 data/
 test/
 wandb/
