+++ ViT/utils/common.py
@@ -21,6 +21,12 @@ COLORS = {
 }
 
 
+def count_parameters(model):
+    ''' Counts number of trainable parameters in the model '''
+
+    return sum([p.numel() for p in model.parameters() if p.requires_grad])
+
+
 def progress_bar(progress=0, status="", bar_len=20):
 
     status = status.ljust(30)
