--- main/configs/bert.yaml
@@ -1,57 +1,0 @@
-
-# Configuration for Bert Self attention
-
-epochs: 300
-batch_size: 100
-eval_every: 10
-
-# Models
-bert_encoder:
-  attention: bert
-  num_heads: 9
-  num_encoder_blocks: 6
-  model_dim: 400
-  ff_dim: 512
-  ff_activation: relu
-  pre_norm: False
-  hierarchical_weight_sharing: True
-
-feature_pooling:
-  resnet: resnet50
-  block: 1
-  pretrained: False
-  model_dim: 400
-  pool_with_resnet: False 
-  pool_downsample_size: 2
-
-clf_head:
-  model_dim: 400
-  num_classes: 10
-
-optimizer:
-  name: sgd
-  lr: 0.1
-  weight_decay: 1.e-04
-
-scheduler:
-  name: cosine
-  warmup_epochs: 15
-
-# Data
-dataset:
-  name: cifar10
-  root: data/cifar10
-  train_transform:
-    random_crop:
-      size: 32
-      padding: 4
-    random_flip:
-    to_tensor:
-    normalize:
-      mean: [0.4914, 0.4822, 0.4465]
-      std: [0.2023, 0.1994, 0.2010]
-  val_transform:
-    to_tensor:
-    normalize:
-      mean: [0.4914, 0.4822, 0.4465]
-      std: [0.2023, 0.1994, 0.2010]
