+++ main/configs/gaussian.yaml
@@ -1,37 +1,37 @@
 
-# Configuration for Bert Self attention
+# Configuration for quadratic encoding setup 
 
 epochs: 300
 batch_size: 32
-eval_every: 10
+eval_every: 10                                                  # Number of training epochs after which validation is done  
 
 # Models
-bert_encoder:
-  attention: gaussian
-  ff_activation: relu
-  num_heads: 16
-  num_encoder_blocks: 6
-  model_dim: 256
-  ff_dim: 400
-  pre_norm: False
-  gaussian_mu_init_noise: 2.0
-  gaussian_sigma_init_noise: 0.01
-  attention_dropout_prob: 0.1
-  attention_gaussian_blur_trick: False
-  attention_isotropic_gaussian: False
-  hierarchical_weight_sharing: False
+encoder:
+  attention: gaussian                                           # Quadratic encoding as the paper calls it 
+  ff_activation: relu                                           # Activation for hidden layer of MLP within attention layer
+  num_encoder_blocks: 6                                         # Number of self-attention layers 
+  num_heads: 16                                                 # Number of heads for self-attention layers
+  model_dim: 256                                                # Working size for the encoder
+  ff_dim: 400                                                   # Size of hidden layer of MLP 
+  pre_norm: False                                               # Whether LayerNorm is before the residual connection (post-norm default)
+  gaussian_mu_init_noise: 2.0                                   # Initialization randomness for attention centers
+  gaussian_sigma_init_noise: 0.01                               # Initialization randomness for attention spreads
+  attention_dropout_prob: 0.1                                   # Dropout prob for attention probs
+  attention_gaussian_blur_trick: False                          # Some way to bypass dot-product attention apparently (NOT USED)
+  attention_isotropic_gaussian: False                           # Whether distro of attention for each head remains isotropic (cI)
+  hierarchical_weight_sharing: False                            # Enable/disable hierarchical self-attention
 
-feature_pooling:
-  resnet: resnet50
-  block: 1
-  pretrained: False
-  model_dim: 256
-  pool_with_resnet: False 
-  pool_downsample_size: 2
+feature_pooling:                                         
+  resnet: resnet50                                              # We are using 2x2 downsampling for pooling, so this is ignored
+  block: 1                                                      # Number of BasicBlocks of resnet to include, if being used
+  pretrained: False                                             # Whether resnet is pretrained
+  model_dim: 256                                                # Working size for the encoder
+  pool_with_resnet: False                                       # Not pooling with resnet by default
+  pool_downsample_size: 2                                       # Image is reversibly downsampled by factor of 2 (default)
 
 clf_head:
-  model_dim: 256
-  num_classes: 10
+  model_dim: 256                                                # in_features into the classification head
+  num_classes: 10                                               # CIFAR-10 has 10 classes
 
 optimizer:
   name: sgd
@@ -43,10 +43,10 @@ scheduler:
   warmup_epochs: 15
 
 # Data
-dataset:
+dataset:                                                        
   name: cifar10
   root: data/cifar10
-  train_transform:
+  train_transform:                                              # Transformations chosen same as in the paper
     random_crop:
       size: 32
       padding: 4
