+++ requirements.txt
@@ -1,0 +1,7 @@
+numpy
+wandb
+Pillow
+datetime
+matplotlib
+torchvision
+cupy-cuda101
